trying to fix unittest
