+++ server/api/crud/files.py
@@ -115,3 +115,18 @@ class Files(
                 err="illegal path prefix or schema",
             )
         return path
+
+    @staticmethod
+    def _verify_and_get_project_secrets(project):
+        # If running on Docker or locally, we cannot retrieve project secrets, so skip.
+        if not server.api.utils.singletons.k8s.get_k8s_helper(
+            silent=True
+        ).is_running_inside_kubernetes_cluster():
+            return {}
+
+        secrets_data = server.api.crud.Secrets().list_project_secrets(
+            project,
+            mlrun.common.schemas.SecretProviderName.kubernetes,
+            allow_secrets_from_k8s=True,
+        )
+        return secrets_data.secrets or {}
