+++ tests/api/api/test_files.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+import unittest.mock
 from http import HTTPStatus
 
 import pytest
@@ -19,6 +20,7 @@ from fastapi.testclient import TestClient
 from sqlalchemy.orm import Session
 
 # fixtures for test, aren't used directly so we need to ignore the lint here
+import mlrun
 from tests.common_fixtures import (  # noqa: F401
     patch_file_forbidden,
     patch_file_not_found,
@@ -44,6 +46,24 @@ def validate_files_status_code(client: TestClient, status_code: int):
 
     resp = client.get("projects/{project}/filestat?schema=v3io&path=mybucket/files.txt")
     assert resp.status_code == status_code
+
+
+class DatastoreObjectMock:
+    def get(self, size, offset):
+        return "dummy body"
+
+    def listdir(self):
+        return ["file1", "file2", "dir1/file3"]
+
+
+@pytest.fixture
+def files_mock():
+    old_object = mlrun.store_manager.object
+    mlrun.store_manager.object = unittest.mock.Mock(return_value=DatastoreObjectMock())
+
+    yield mlrun.store_manager.object
+
+    mlrun.store_manager.object = old_object
 
 
 def test_files(db: Session, client: TestClient, files_mock, k8s_secrets_mock) -> None:
