+++ tests/api/crud/test_files.py
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+import unittest.mock
+
 import fastapi.testclient
 import sqlalchemy.orm
 
@@ -19,10 +21,9 @@ import mlrun.common.schemas
 import server.api.crud
 
 
-def test_delete_atifact_data(
+def test_delete_artifact_data(
     db: sqlalchemy.orm.Session,
     client: fastapi.testclient.TestClient,
-    files_mock,
     k8s_secrets_mock,
 ) -> None:
     path = "s3://somebucket/some/path/file"
@@ -33,12 +34,15 @@ def test_delete_atifact_data(
     full_secrets = project_secrets.copy()
     full_secrets.update(env_secrets)
     k8s_secrets_mock.store_project_secrets(project, project_secrets)
+    mlrun.datastore.store_manager.object = unittest.mock.Mock()
 
     server.api.crud.Files().delete_artifact_data(
         mlrun.common.schemas.AuthInfo(), project, path
     )
-    files_mock.assert_called_once_with(url=path, secrets=full_secrets, project=project)
-    files_mock.reset_mock()
+    mlrun.datastore.store_manager.object.assert_called_once_with(
+        url=path, secrets=full_secrets, project=project
+    )
+    mlrun.datastore.store_manager.object.reset_mock()
 
     # user supplied secrets - use the same key to override project secrets
     user_secrets = {"secret1": "user-secret"}
@@ -47,7 +51,7 @@ def test_delete_atifact_data(
     server.api.crud.Files().delete_artifact_data(
         mlrun.common.schemas.AuthInfo(), project, path, secrets=user_secrets
     )
-    files_mock.assert_called_once_with(
+    mlrun.datastore.store_manager.object.assert_called_once_with(
         url=path, secrets=override_secrets, project=project
     )
-    files_mock.reset_mock()
+    mlrun.datastore.store_manager.object.reset_mock()
